chore:some UI Improvements
